% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoms.R
\name{get_geoms}
\alias{get_geoms}
\title{Get spatial polygons for specified geography}
\usage{
get_geoms(geography, state)
}
\arguments{
\item{geography}{A dataframe}

\item{state}{A variable in the dataframe}
}
\value{
The dataframe with new mean and sum columns
}
\description{
Get spatial polygons for specified geography
}
