% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sovi.R
\name{get_sovi}
\alias{get_sovi}
\title{Get specified social vulnerability variables from the census}
\usage{
get_sovi(
  geography,
  state = NULL,
  variables = NULL,
  year = 2015,
  geometry = FALSE
)
}
\arguments{
\item{geography}{A dataframe}

\item{state}{A variable in the dataframe}

\item{variables}{A variable in the dataframe}

\item{year}{A variable in the dataframe}

\item{geometry}{A variable in the dataframe}
}
\value{
The dataframe with new mean and sum columns
}
\description{
Get specified social vulnerability variables from the census
}
